requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pillow==10.4.0
requests==2.32.3
soupsieve==2.6
urllib3==2.2.2
